requirements.txt
asgiref==3.9.1
Django==5.0.2
django-cors-headers==4.3.1
django-unfold==0.21.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
